Format and cleanup comments
